seeding issue fixed
The problem with my initial seeding method was that I was seeding with a valid control vect but the actor was acting independently. Now the actor gets trained to generate the initial seed. also high variance (small variance in control space means very large variance in flux space) was causing high penalty in infeasibility. setting the actor networks variance to a lower value worked for now